Implement augmented assignment for misc types.
I believe that to be all of the types in the repository, so anything
where this works (for any of the operators: `+`, `-`, `/`, `*`, `%`,
`^`, `|`, `&`, `<<`, `>>`):

```rust
a = a + b;
```

This will also now work:

```rust
a += b;
```

(Well, when the langitem is implemented, anyway.)

AND, what's more, it may be more efficient (e.g. for a vector it won't
create a new vector, it will just push the items onto self).

Not tested in the slightest bit yet. That will come.
(Building is much too slow on my machine.)