Make it compile again.
There were various basic errors in what I did earlier but I hadn't tried
compiling it...

Also notable at this point is the addition of an implementation of Add
for ~str which was not there before; automatic casting to an &str took
care of letting + work for it. I have assurance from dbaupp (@huonw)
that it is OK to add this.

Naturally, this being here still doesn't mean that it *works*.