Get augmented operators to almost work.
Currently any attempt to use an augmented assignment operator which
should be valid (i.e. the *Assign trait is implemented correctly) is
failing in an LLVM assertion as it seems an extra argument is being
prepended to the function call. Still trying to figure this one out; I'm
rather out of my depth. :-)